anchor/pyproject.toml
[project]
name = "anchor"
version = "0.1.0"
description = "Anchor — a companion that holds memory for people who are losing theirs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "openai>=1.14.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "pydantic>=2.6.0",
]